Add files via upload
Added 5 files. Head.cpp which opens and returns the first 10 lines of all files passed via command line arguments. Tail which does the same, but with the final 10 lines. PWD, which prints the current working directory. WC which opens the given files and counts the number of lines, word, and bytes. Then also calculates a total a the end (of all given CLA files).